chnges in code

+++ chp6_conditio nal/nested if.py
@@ -1,17 +1,17 @@
 # nested if 
-cgpa = int(input("enter your cgpa : "))
+cgpa = float(input("enter your cgpa : "))
 if(cgpa>=7.5):
     print("congs you are eligible :  ")
 
-    b = int(input("enter your pending backlog :"))
+    b = float(input("enter your pending backlog :"))
     if(b==0):
         print("well done go ahead : ")
 
-        c = int(input("enter your 10th percentage : "))
+        c = float(input("enter your 10th percentage : "))
         if(c >= 85):
             print("great")
 
-            d= int(input("enter +2 percenatge :"))
+            d= float(input("enter +2 percenatge :"))
 
             if(d>85):
                 print("you are eligible for job : ")
